// web/src/api/models/UpdateCredentialsRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * mindscape
 * This is the swagger page for the Project mindscape generated with Egg-go. use this to test your database connection
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateCredentialsRequest
 */
export interface UpdateCredentialsRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateCredentialsRequest
     */
    email?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCredentialsRequest
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCredentialsRequest
     */
    oldPassword?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCredentialsRequest
     */
    password?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateCredentialsRequest
     */
    username?: string;
}

/**
 * Check if a given object implements the UpdateCredentialsRequest interface.
 */
export function instanceOfUpdateCredentialsRequest(value: object): value is UpdateCredentialsRequest {
    return true;
}

export function UpdateCredentialsRequestFromJSON(json: any): UpdateCredentialsRequest {
    return UpdateCredentialsRequestFromJSONTyped(json, false);
}

export function UpdateCredentialsRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateCredentialsRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'email': json['email'] == null ? undefined : json['email'],
        'id': json['id'] == null ? undefined : json['id'],
        'oldPassword': json['oldPassword'] == null ? undefined : json['oldPassword'],
        'password': json['password'] == null ? undefined : json['password'],
        'username': json['username'] == null ? undefined : json['username'],
    };
}

export function UpdateCredentialsRequestToJSON(json: any): UpdateCredentialsRequest {
    return UpdateCredentialsRequestToJSONTyped(json, false);
}

export function UpdateCredentialsRequestToJSONTyped(value?: UpdateCredentialsRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'email': value['email'],
        'id': value['id'],
        'oldPassword': value['oldPassword'],
        'password': value['password'],
        'username': value['username'],
    };
}